requirements.txt
django==1.9.5
django-flat-theme
easy_thumbnails==2.3
django-image-cropping==1.0.3
django-tagging==0.4.1
django-annoying==0.8.1
django-contrib-comments==1.7.0
django-decouple==2.1
django-widget-tweaks==1.3
django-ckeditor==4.4.4
django-taggit==0.15.0